Attempt to fix install error
When installing the RPM, if it has already been installed, makes a copy
of the keys. After all the installation, it moves them back to the
original name. If it's the first time though, it always tries to move
them back even they don't exist. This commit tries to fix this issue.